support make install from the source repo
This was supported from the tarball due to the man page
being distributed.  However man/dvd-vr.1 needs to be
created when building from the source repo.
Rather than adding a dependence on help2man and perl
for anyone that might be interacting with the repo,
we provide a man page to be copied into place when
there is no existing man/dvd-vr.1 file present.