Fix Makefile checks for iconv
Either the C preprocessor seems to no longer accept '\#' as '#' or GNU Make
or the shell command 'echo' no longer transform '\#' into '#', as a result,
gcc reports:
<stdin>:1:1: error: stray ‘\’ in program
This compiler error masks the intention of the tests, so I have decided to
remove the backslash. Moreover, cpp does not work unless being invoked with
option '-xc'.